Factor out `ArraySlice` into its own TypeScript goody
I would like to use the `ArraySlice` class to do some work with combinations, permutations, and subsets. It will be nice to have it as a separate goody for this purpose.
